+++ akita/src/api/store.ts
@@ -10,6 +10,7 @@ import { __globalState } from '../internal/global-state';
 import { getAkitaConfig } from './config';
 
 let __DEV__ = true;
+const isNotBrowser = typeof window === 'undefined';
 
 export const __stores__: { [storeName: string]: Store<any> } = {};
 
@@ -255,7 +256,8 @@ export class Store<S> {
   }
 
   ngOnDestroy() {
-    if( !(window as any).hmrEnabled && this === __stores__[this.storeName] ) {
+    if (isNotBrowser) return;
+    if (!(window as any).hmrEnabled && this === __stores__[this.storeName]) {
       delete __stores__[this.storeName];
     }
   }