+++ akita/src/api/types.ts
@@ -47,3 +47,5 @@ export type Newable<T> = { new (...args: any[]): T };
 
 /** Optional config for Store.add() to prepend entity or entities to stores */
 export type AddOptions = { prepend?: boolean };
+
+export type SetActiveOptions = { prev?: boolean; next?: boolean; wrap?: boolean };