+++ akita/src/plugins/filters/filters-utils.ts
@@ -4,11 +4,11 @@ import { Filter } from './filters-store';
 /**
  * Helper function to do a default filter
  */
-export function defaultFilter<E = any>( inElement: E, index: number, array: E[], filter: Filter<E> ): boolean {
-  if( isObject(inElement) && isDefined(filter.value) && isString(filter.value) ) {
-    return searchFilter(filter.value, inElement);
+export function defaultFilter<E = any>( value: E, index: number, array: E[], filter: Filter<E> ): boolean {
+  if( isObject(value) && isString(filter.value) ) {
+    return searchFilter(filter.value, value);
   }
-  return isDefined(filter.value) ? filter.value === inElement : !!inElement;
+  return isDefined(filter.value) ? filter.value === value : !!value;
 }
 
 /**
@@ -24,5 +24,5 @@ export function searchFilter( searchKey: string, inObj: Object ): boolean {
  * Helper function to do search in one key of an object
  */
 export function searchFilterIn( searchKey: string, inObj: Object, inKey: string ): boolean {
-  return typeof inObj[inKey] === 'string' && inObj[inKey].toLocaleLowerCase().includes(searchKey.toLocaleLowerCase());
+  return isString(inObj[inKey]) && inObj[inKey].toLocaleLowerCase().includes(searchKey.toLocaleLowerCase());
 }