+++ playground/src/app/nav/nav.component.ts
@@ -40,5 +40,6 @@ export class NavComponent {
 
   resetStores() {
     resetStores();
+    this.router.navigateByUrl('login');
   }
 }